requirements.txt
amqp==5.1.1
billiard==3.6.4.0
blinker==1.6.2
CacheControl==0.12.11
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.2
Deprecated==1.2.13
firebase-admin==6.1.0
Flask==2.3.1
Flask-Cors==3.0.10
Flask-Limiter==3.3.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
httplib2==0.22.0
idna==3.4
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
limits==3.4.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
msgpack==1.0.5
ordered-set==4.1.0
packaging==23.1
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==1.0.0
pytz==2023.3
requests==2.29.0
rich==13.3.4
rsa==4.9
six==1.16.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.0
wrapt==1.15.0
